Added allow_hidden to is_product_installable and install_product.
Default False, which is what is used in the add-ons control panel.

When True, hidden profiles and products (marked as INonInstallable)
can be installed anyway.  This is meant to be used from code where you
explicitly want to install this.  Otherwise core Plone packages that
are hidden from the user, cannot be installed.

After all, if you *really* do not want a profile to be installable
ever, then you should remove it.